+++ tests/testthat/test-coerce-aggregation-structure.R
@@ -11,6 +11,7 @@ test_that("coercion methods works", {
            dimnames = list(c("1", "2", "11", "21", "12"),
                            c("111", "211", "121", "112")))
   )
+  expect_equal(as.matrix(agg), as.matrix(as.matrix(agg, sparse = TRUE)))
   expect_equal(
     as.data.frame(agg),
     data.frame(level1 = c("1", "2", "1", "1"),
@@ -21,23 +22,23 @@ test_that("coercion methods works", {
   expect_equal(agg, as_aggregation_structure(as.data.frame(agg)))
   expect_equal(as.list(agg), list(x1, x2, x3))
 
-  expect_equal(as.matrix(aggregation_structure(1:3, 2)),
+  agg2 <- aggregation_structure(1:3, 2)
+  expect_equal(as.matrix(agg2),
                matrix(c(1, 1), 2, 1, dimnames = list(c("1", "2"), "3")))
-  expect_equal(as.data.frame(aggregation_structure(1:3, 2)),
+  expect_equal(as.matrix(agg2), as.matrix(as.matrix(agg2, sparse = TRUE)))
+  expect_equal(as.data.frame(agg2),
                data.frame(level1 = "1", level2 = "2", ea = "3", weight = 2))
-  expect_equal(as.list(aggregation_structure(1:3, 2)),
-               as.list(as.character(1:3)))
+  expect_equal(as.list(agg2), as.list(as.character(1:3)))
 
+  agg3 <- aggregation_structure(list(letters[c(2, 1, 3)]), 1:3)
   expect_equal(
-    as.matrix(aggregation_structure(list(letters[c(2, 1, 3)]), 1:3)),
+    as.matrix(agg3),
     matrix(numeric(0), 0, 3, dimnames = list(character(0), c("b", "a", "c")))
   )
+  expect_equal(as.matrix(agg3), as.matrix(as.matrix(agg3, sparse = TRUE)))
   expect_equal(
-    as.data.frame(aggregation_structure(list(letters[c(2, 1, 3)]), 1:3)),
+    as.data.frame(agg3),
     data.frame(ea = c("b", "a", "c"), weight = 1:3)
   )
-  expect_equal(
-    as.list(aggregation_structure(list(letters[c(2, 1, 3)]), 1:3)),
-    list(letters[c(2, 1, 3)])
-  )
+  expect_equal(as.list(agg3), list(letters[c(2, 1, 3)]))
 })